% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{applyNS}
\alias{applyNS}
\title{applyNS
Counts number of trades during last k seconds}
\usage{
applyNS(s, k = 1.5)
}
\arguments{
\item{s}{time-vector (seconds)}

\item{k}{(default = 1.5)}
}
\value{
cnt vector of same length at s, indicating number of occurences during [t,t-k]
}
