% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autocovariance}
\alias{autocovariance}
\title{Autocovariane}
\usage{
autocovariance(rftdata, hval)
}
\description{
computes the Autocovariance matrices
inputs are Refresh time sampled returns and optimal Hval
}
