% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_estimates.R
\name{db_estimates}
\alias{db_estimates}
\title{Drift-Burst t-statistic
data(lobster)
lobster}
\usage{
db_estimates(lobster, testing.time = unique(floor(data$Secs/2) * 2))
}
\description{
Drift-Burst t-statistic
data(lobster)
lobster
}
