% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftburst_tstat.R
\name{driftburst_tstat}
\alias{driftburst_tstat}
\title{Drift-Burst t-statistic
Function computes drift-burst statistics from [Christensen, Oomen, Reno](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2842535)}
\usage{
driftburst_tstat(lobster, testing.time = lobster$Secs[-c(1:5)])
}
\arguments{
\item{lobster}{Data-frame from lobster}

\item{testing.time}{Testing time (default = every observations (after discarding the first 5))}
}
\value{
vector with t stat values
}
