% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalue_cleaning.R
\name{eigenval_clean}
\alias{eigenval_clean}
\title{eigenvalue cleaning following Hautsch,Kyj, Oomen 2010}
\usage{
eigenval_clean(cov_org)
}
\description{
eigenvalue cleaning following Hautsch,Kyj, Oomen 2010
}
