% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exponential.kernel}
\alias{exponential.kernel}
\title{Exponential kernel}
\usage{
exponential.kernel(x)
}
\arguments{
\item{x}{input value (vectors also allowed)}
}
\value{
Exponential kernel evaluated at x
}
