% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_LOBSTER_files.R
\name{extract_LOBSTER_files}
\alias{extract_LOBSTER_files}
\title{extract_LOBSTER_files}
\usage{
extract_LOBSTER_files(date, ticker, startdate = toDate(2007, 6, 27),
  enddate = toDate(2016, 6, 27), starttrade = 9.5 * 60 * 60, endtrade = 16
  * 60 * 60, nlevels = 1)
}
\description{
This function extracts 7z files containing LOBSTER date (NOTE: 7z has to be included into the system PATH!)
}
