% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IVhat_f}
\alias{IVhat_f}
\title{IV estimate based on 20 minute RV estimates}
\usage{
IVhat_f(rftprices)
}
\arguments{
\item{rftprices}{Refresh time sampled (log) prices (xts object with N columns (assets) and T rows (observations))}
}
\description{
IV estimate based on 20 minute RV estimates
}
