% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobster.R
\docType{data}
\name{lobster}
\alias{lobster}
\title{Lobster: Limit order book sample}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.lobsterdata.com}
}
\usage{
lobster
}
\description{
Amazond data sample
}
\keyword{datasets}
