% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parzen.kernel}
\alias{parzen.kernel}
\title{Parzen Kernel
Computes the Parzen-Kernel}
\usage{
parzen.kernel(x)
}
\arguments{
\item{x}{value (also vectors allowed) at which to evaluate the Parzen-Kernel}
}
\value{
w Parzen-Kernel
}
