% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_lobster.r
\name{readin_lobster}
\alias{readin_lobster}
\title{readin_lobster}
\usage{
readin_lobster(ticker, date, nlevels = 1, output = "ALL")
}
\description{
This function extracts 7z files containing LOBSTER date (NOTE: 7z has to be included into the system PATH!)
}
