% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realized_kernel.R
\name{realized_kernel}
\alias{realized_kernel}
\title{Realized Kernel estimator (univariate)}
\usage{
realized_kernel(hft_prices)
}
\arguments{
\item{hft_prices}{vector with (log) prices}
}
\description{
Realized Kernel estimator (univariate)
}
