% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_lobster.R
\name{remove_lobster}
\alias{remove_lobster}
\title{remove_lobster}
\usage{
remove_lobster(date, ticker, nlevels = 1)
}
\description{
This function extracts 7z files containing LOBSTER date (NOTE: 7z has to be included into the system PATH!)
}
