% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_scale_realized_volatility.R
\name{two_scale_realized_volatility}
\alias{two_scale_realized_volatility}
\title{Two-scaled realized volatility estimator}
\usage{
two_scale_realized_volatility(logprices, K = 300, J = 1)
}
\arguments{
\item{logprices}{vector with log prices}

\item{K}{(Slow) Time Scale (in Seconds), default is 300 Seconds (=5 Minutes)}

\item{J}{(Fast) Time Scale (in Seconds), default is 1 Second}
}
\value{
Realized volatility estimator
}
