% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baing.R
\name{baing_clean}
\alias{baing_clean}
\title{regularization based on factor structure following Hautsch,Kyj, Oomen 2010}
\usage{
baing_clean(cov_org, factors = 3)
}
\description{
regularization based on factor structure following Hautsch,Kyj, Oomen 2010
}
