% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_estimate.R
\name{brk_estimate}
\alias{brk_estimate}
\title{BRK Estimate
ticker <- as.character(read.table('AddInformation/assets.csv')$x)
sim_days <- as.Date(read.table('AddInformation/trading_days.csv')$x)}
\usage{
brk_estimate(tickerlist, date, nob = 6, folder = ".")
}
\description{
BRK Estimate
ticker <- as.character(read.table('AddInformation/assets.csv')$x)
sim_days <- as.Date(read.table('AddInformation/trading_days.csv')$x)
}
