% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalue_cleaning.R
\name{eigenval_clean}
\alias{eigenval_clean}
\title{Eigenvalue cleaning following Hautsch, Kyj, Oomen 2010}
\usage{
eigenval_clean(cov_org, n_obs = 2000)
}
\arguments{
\item{cov_org}{Variance-Covariance Matrix you want to clean}

\item{n_obs}{Number of observations used to estimate the matrix (default = 2000)}
}
\value{
Cleaned Matrix of same size as cov_org
}
