% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lobster.R
\name{extract_lobster}
\alias{extract_lobster}
\title{extract_lobster}
\usage{
extract_lobster(date, ticker, nlevels = 1)
}
\description{
This function extracts 7z files containing LOBSTER date (NOTE: 7z has to be included into the system PATH!)
}
