% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobster.R
\docType{data}
\name{lobster}
\alias{lobster}
\title{Lobster: Limit order book sample}
\format{A data frame with 53.940 rows and 10 variables:
\describe{
  \item{Secs}{Second since midnight}
  ...
}}
\source{
\url{http://www.lobsterdata.com}
}
\usage{
lobster
}
\description{
Amazon data sample
}
\keyword{datasets}
