% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_lobster.R
\name{readin_lobster}
\alias{readin_lobster}
\title{readin_lobster}
\usage{
readin_lobster(ticker, date, nlevels = 1, output = "ALL", folder = ".")
}
\description{
This function reads-in extracted Lobster Files (either Orderbook, Messagebook or both)
}
